Fix alignment issues
Storing the data in a `u128` was not sound. This is because the
alignment of `Plid` was 16 but we only told Postgres to use 4 (via
implicit `ALIGNMENT` in CREATE TYPE). Even if we did bump this,
Postgres only supports 8 at most. This would mean that Postgres would
allocate values with incorrect alignment.

On x86_64 in particular this seg faults with EXC_BAD_ACCESS (code=EXC_I386_GPFLT).